chore(feat): Reorganized file structure and tests, fixed typo
